feat: prepared another version of data table
